main: extract event dispatcher and shutdown plumbing
cef_consumer_thread, the mpv event queue, and update_idle_inhibit move to
event_dispatcher.{h,cpp} — the mpv→browsers/MediaSession bridge is now in
one place instead of strewn across main's globals. g_was_maximized_before_
fullscreen moves with it (writer is the consumer thread).

g_shutting_down, g_shutdown_event, initiate_shutdown, signal_handler, and
the try_close_browser helper move to shutdown.{h,cpp}. common.h's existing
forward decls keep all current includers compiling unchanged.

main.cpp drops from 1015 to 820 lines; event_queue.h, wake_event.h, and
cef_parser.h are no longer needed there.
